moved ReaderSearchService to a new package /reader/services/search
