+++ WordPress/src/main/java/org/wordpress/android/ui/reader/services/search/ReaderSearchService.java
@@ -1,4 +1,4 @@
-package org.wordpress.android.ui.reader.services;
+package org.wordpress.android.ui.reader.services.search;
 
 import android.app.Service;
 import android.content.Context;
